+++ src/main/java/roidrole/thaumicinfo/hwyla/ProviderGogglesDisplay.java
@@ -5,7 +5,7 @@ import mcp.mobius.waila.api.IWailaDataAccessor;
 import mcp.mobius.waila.api.IWailaDataProvider;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
-import roidrole.thaumicinfo.Tags;
+import roidrole.thaumicinfo.ThaumicInformationConfig;
 import thaumcraft.api.items.IGogglesDisplayExtended;
 import thaumcraft.common.lib.utils.EntityUtils;
 
@@ -25,7 +25,7 @@ public class ProviderGogglesDisplay implements IWailaDataProvider {
 		if (!(te instanceof IGogglesDisplayExtended)) {
 			return tooltip;
 		}
-		if (config.getConfig(Tags.MOD_ID+".require_goggles") && !EntityUtils.hasGoggles(accessor.getPlayer())) {
+		if (ThaumicInformationConfig.hwylaConfig.requireGoggles && !EntityUtils.hasGoggles(accessor.getPlayer())) {
 			return tooltip;
 		}